+++ package.json
@@ -34,10 +34,6 @@
     "@as-integrations/express5": "1.1.2",
     "@aws-sdk/client-s3": "3.966.0",
     "@aws-sdk/s3-request-presigner": "3.966.0",
-    "@babel/core": "7.28.5",
-    "@babel/node": "7.28.0",
-    "@babel/plugin-transform-typescript": "7.28.5",
-    "@babel/preset-env": "7.28.5",
     "@escape.tech/graphql-armor": "3.2.0",
     "@hyperwatch/hyperwatch": "4.2.0",
     "@json2csv/plainjs": "^7.0.6",
@@ -59,7 +55,6 @@
     "argparse": "2.0.1",
     "async-mutex": "0.5.0",
     "axios": "1.13.2",
-    "babel-plugin-add-module-exports": "1.0.4",
     "bayes": "1.0.0",
     "bcrypt": "6.0.0",
     "blurhash": "2.0.5",
@@ -137,8 +132,6 @@
     "zod": "4.3.6"
   },
   "devDependencies": {
-    "@babel/cli": "^7.23.4",
-    "@babel/register": "^7.23.7",
     "@graphql-eslint/eslint-plugin": "^4.0.0",
     "@opentelemetry/api": "^1.7.0",
     "@opentelemetry/auto-instrumentations-node": "^0.67.0",
@@ -155,7 +148,6 @@
     "@types/nodemailer": "^7.0.4",
     "@types/validator": "^13.11.8",
     "@typescript-eslint/parser": "^8.13.0",
-    "babel-plugin-istanbul": "^7.0.1",
     "chai": "^6.0.0",
     "chai-as-promised": "^7.1.1",
     "chai-jest-snapshot": "^2.0.0",
@@ -192,6 +184,7 @@
     "sinon": "^21.0.0",
     "sinon-chai": "^3.7.0",
     "supertest": "^7.0.0",
+    "ts-node": "10.9.2",
     "ts-unused-exports": "^11.0.1",
     "typescript": "5.9.3",
     "typescript-eslint": "^8.13.0",
@@ -200,11 +193,11 @@
   "scripts": {
     "build": "npm run build:clean && npm run build:updates && npm run build:server",
     "build:clean": "rm -rf dist && mkdir dist",
-    "build:server": "babel server --copy-files --extensions .js,.ts -d ./dist",
+    "build:server": "tsc -p tsconfig.build.json && rsync -a --exclude '*.ts' --exclude '*.js' server/ dist/",
     "build:updates": "npm --prefix node_modules/cloudflare-ip run update-list",
     "clean": "npm run build:clean && npm run test:clean",
     "commit": "git-cz",
-    "compile:email": "babel-node --extensions .js,.ts scripts/compile-email.js",
+    "compile:email": "npm run script -- scripts/compile-email.js",
     "compile:email:watch": "./scripts/watch_email_template.sh",
     "cron:10mn": "./scripts/cron.sh 10mn",
     "cron:daily": "./scripts/cron.sh daily",
@@ -212,25 +205,25 @@
     "cron:monthly": "./scripts/cron.sh monthly",
     "cron:weekly": "./scripts/cron.sh weekly",
     "cron:yearly": "./scripts/cron.sh yearly",
-    "db:clean": "babel-node --extensions .js,.ts ./scripts/sanitize-db.ts",
+    "db:clean": "npm run script -- ./scripts/sanitize-db.ts",
     "db:copy:staging": "./scripts/db_copy_to_localhost.sh staging",
     "db:dump:dev": "./scripts/db_migrate_opencollective_dvl.sh",
     "db:migrate": "./scripts/sequelize.sh db:migrate",
     "db:migrate:undo": "./scripts/sequelize.sh db:migrate:undo",
     "db:migration:create": "./scripts/sequelize.sh migration:create",
-    "db:migration:archive": "babel-node --extensions .js,.ts scripts/db_migrations_archive.ts",
+    "db:migration:archive": "npm run script -- scripts/db_migrations_archive.ts",
     "db:restore": "./scripts/db_restore.sh",
     "db:restore:e2e": "npm run db:restore -- -d opencollective_e2e",
     "db:restore:test": "NODE_ENV=test npm run db:restore -- -d opencollective_test && NODE_ENV=test npm run db:migrate",
-    "db:sanitize": "babel-node --extensions .js,.ts scripts/sanitize-db.ts",
-    "db:sanitize-collective-tags": "babel-node --extensions .js,.ts scripts/sanitize-collective-tags",
+    "db:sanitize": "npm run script -- scripts/sanitize-db.ts",
+    "db:sanitize-collective-tags": "npm run script -- scripts/sanitize-collective-tags",
     "db:seed": "./scripts/sequelize.sh db:seed",
-    "db:setup": "babel-node --extensions .js,.ts ./scripts/db_setup.js",
+    "db:setup": "npm run script -- ./scripts/db_setup.js",
     "depcheck": "npx @opencollective/depcheck .",
     "deploy:production": "./scripts/pre-deploy.sh production && git push production main",
     "deploy:staging": "./scripts/pre-deploy.sh staging && git push -f staging main",
-    "dev": "nodemon server/index.ts -x \"babel-node --extensions .js,.ts\" . -e js,hbs,ts",
-    "docs:generate:oauth": "babel-node --extensions .js,.ts ./scripts/docs/oauth.ts",
+    "dev": "nodemon server/index.ts -x \"npm run script --\" . -e js,hbs,ts",
+    "docs:generate:oauth": "npm run script -- ./scripts/docs/oauth.ts",
     "graphql:update": "npm-run-all graphql:updateV1 graphql:updateV2",
     "graphql:updateV1": "npm run script scripts/update-gql-schemas.ts http://localhost:3060/graphql/v1 server/graphql/schemaV1.graphql && prettier server/graphql/schemaV1.graphql --write",
     "graphql:updateV2": "npm run script scripts/update-gql-schemas.ts http://localhost:3060/graphql/v2 server/graphql/schemaV2.graphql && prettier server/graphql/schemaV2.graphql --write",
@@ -244,10 +237,10 @@
     "prettier": "prettier \"**/*.@(js|ts|json|md)\"",
     "prettier:check": "npm run prettier -- --check",
     "prettier:write": "npm run prettier -- --write",
-    "script": "babel-node --extensions .js,.ts $1",
+    "script": "TS_NODE_FILES=true node -r ts-node/register $1",
     "search": "./scripts/dev/run-docker-compose.sh -f docker-compose/search.yml up",
     "minio": "./scripts/dev/run-docker.sh compose -f docker-compose/uploads.yml up",
-    "sequelize": "babel-node --extensions .js,.ts -- ./node_modules/.bin/sequelize --config config/sequelize-cli.js",
+    "sequelize": "npm run script -- ./node_modules/.bin/sequelize --config config/sequelize-cli.js",
     "start": "node ./dist/index.js",
     "start:e2e": "TZ=UTC NODE_ENV=production OC_ENV=e2e E2E_TEST=1 npm-run-all db:restore:e2e db:migrate build start:e2e:server",
     "start:e2e:server": "node ./dist/e2e.js",
@@ -296,7 +289,6 @@
       "migrations/archives"
     ],
     "specials": [
-      "babel",
       "bin",
       "commitizen",
       "eslint",
