+++ scripts/cron.sh
@@ -1,11 +1,8 @@
 #!/bin/bash
-BIN_PATH=./node_modules/.bin
-BIN=babel-node
-BIN_OPTIONS="--extensions .js,.ts"
 FILE_PATH="./cron/$1/"
 TZ=utc
 
 for FILE in `ls $FILE_PATH`; do 
-  echo "Running $BIN $FILE_PATH$FILE";
-  "$BIN_PATH/$BIN" ${BIN_OPTIONS} -r "@sentry/node/preload" "$FILE_PATH$FILE";
+  echo "Running cron job $FILE_PATH$FILE";
+  NODE_OPTIONS="-r @sentry/node/preload" npm run script -- "$FILE_PATH$FILE";
 done;
