+++ scripts/sequelize.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# This script wraps the sequelize command with babel-node and passes
+# This script wraps the sequelize command with npm run script and passes
 # parameters required in every single call.
 
 # Important paths
@@ -9,7 +9,7 @@ PATH=${PATH}:$NODEBIN
 
 # Parameters & Command
 SEQUELIZE_CONFIG="--models-path server/models/ --config config/sequelize-cli.js"
-COMMAND="babel-node --extensions .js,.ts $NODEBIN/sequelize ${SEQUELIZE_CONFIG} $@"
+COMMAND="npm run script -- $NODEBIN/sequelize ${SEQUELIZE_CONFIG} $@"
 
 # Variables exported for the exec
 export CUSTOM_DB PATH
