--- test/setup-babel.cjs
@@ -1,1 +1,0 @@
-require('@babel/register')({ extensions: ['.js', '.ts'] });
